pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.12.3",
    "google-cloud-aiplatform>=1.116.0",
    "ipykernel>=6.30.1",
    "pyyaml>=6.0.2",
]
